add failing test for preemptive interruptions
